Add gulp dts task which generates single .d.ts from multiple files as workaround.
This uses concatenation which means we can use default exports.
